requirements.txt
Bottleneck==1.3.4
brotlipy==0.7.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.4
cryptography==37.0.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
pip==21.2.4
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
setuptools==61.2.0
six==1.16.0
SQLAlchemy==1.4.36
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2